src/main/java/io/artur/spring/tutorial/mvc/sfgrecipe/services/RecipeServiceImpl.java
package io.artur.spring.tutorial.mvc.sfgrecipe.services;

import io.artur.spring.tutorial.mvc.sfgrecipe.domain.Recipe;
import io.artur.spring.tutorial.mvc.sfgrecipe.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.HashSet;
import java.util.Optional;
import java.util.Set;

/**
 *
 */
@Slf4j
@Service
public class RecipeServiceImpl implements RecipeService{

    private final RecipeRepository recipeRepository;

    public RecipeServiceImpl(RecipeRepository recipeRepository) {
        this.recipeRepository = recipeRepository;
    }

    @Override
    public Set<Recipe> getRecipes() {

        log.debug("Retrieving recipes");
        Set<Recipe> recipes = new HashSet<>();

        recipeRepository.findAll().iterator().forEachRemaining(recipes::add);

        return recipes;
    }

    @Override
    public Recipe findById(Long id) {
        Optional<Recipe> optionalRecipe = recipeRepository.findById(id);
        if (optionalRecipe.isEmpty()) {
            throw new RuntimeException("Recipe not found!");
        }
        return optionalRecipe.get();
    }
}
